Rename ModClientHandler to ClientHandler
