+++ src/main/java/com/fredtargaryen/rocketsquids/client/render/RenderBabyRS.java
@@ -13,7 +13,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 import org.jetbrains.annotations.NotNull;
 
-import static com.fredtargaryen.rocketsquids.client.event.ModClientHandler.BABY_SQUID_BODY_LAYER;
+import static com.fredtargaryen.rocketsquids.client.event.ClientHandler.BABY_SQUID_BODY_LAYER;
 
 @SuppressWarnings("removal")
 public class RenderBabyRS extends MobRenderer<BabyRocketSquidEntity, ModelRocketSquidBaby<BabyRocketSquidEntity>> {
