+++ src/main/java/com/fredtargaryen/rocketsquids/network/message/MessageSquidNote.java
@@ -2,7 +2,7 @@
 // See README.md for full copyright notice and contributor info
 package com.fredtargaryen.rocketsquids.network.message;
 
-import com.fredtargaryen.rocketsquids.client.event.ModClientHandler;
+import com.fredtargaryen.rocketsquids.client.event.ClientHandler;
 import io.netty.buffer.ByteBuf;
 import net.minecraftforge.network.NetworkEvent;
 
@@ -27,7 +27,7 @@ public class MessageSquidNote {
 
     public void onMessage(Supplier<NetworkEvent.Context> ctx) {
         ctx.get().enqueueWork(() -> {
-            ModClientHandler.playNoteFromMessageConchNeeded(this.note);
+            ClientHandler.playNoteFromMessageConchNeeded(this.note);
         });
         ctx.get().setPacketHandled(true);
     }
